+++ CHANGELOG.md
@@ -7,6 +7,16 @@ Versioning follows [Semantic Versioning](https://semver.org/).
 
 ---
 
+## [0.5.0-beta] - 2026-02-24
+
+### Added
+- `gui/script_tab.py` : onglet Script
+
+### Changed
+- `gui/main_window.py` : replacement placeholder Script by ScriptTab
+
+---
+
 ## [0.4.0-beta] - 2026-02-24
 
 ### Added